+++ Userland/Libraries/LibGL/CMakeLists.txt
@@ -1,18 +1,6 @@
 set(SOURCES
-    GLBlend.cpp
-    GLColor.cpp
+    GLAPI.cpp
     GLContext.cpp
-    GLDraw.cpp
-    GLFog.cpp
-    GLLights.cpp
-    GLLists.cpp
-    GLMap.cpp
-    GLMatrix.cpp
-    GLStencil.cpp
-    GLTexture.cpp
-    GLUtils.cpp
-    GLVert.cpp
-    GLVertexArrays.cpp
     Tex/NameAllocator.cpp
     Tex/Texture2D.cpp
 )
